dedcode: this doesn't have an impact.
I'm reading docs which seem to indicate this function must be called before linking the program, but the example I'm coming from doesn't, and I've yet to find any explanation for what this is *supposed* to do, so, I'm dropping it.

yes, that makes one wonder what that output in that shaders is going into.  But clearly something (else); if you comment out the fragment shader from the program, you get a black cube.