Comments.
Have spent a bunch of time theorycrafting about how to best make an interface that lets game objects render themselves as reasonable atoms, but it's a messy question with different design goals pulling in different directions... I just don't know enough to wrap a good API around all this stuff yet.  So, that's sitting on the sidelines, and I'm going to go back to more ginger proding.