Start pulling a reboot on our gl stuff.
We can probably do something better now, between the learning we've had about GL, and the availability of go-sup and meep for good concurrent actor and error handling defitions, both of which should be nice bonus leverage.

"egl" because ericGL.  fuck the police.  (or more constructively: i'm under no illusions about trying to make this extractable and reusable to anyone else.)

Signed-off-by: Eric Myhre <hash@exultant.us>